requirements.txt
yaml
pillow
TeraTTS
ruaccent
google-generativeai
io
transliterate